fix: set up for backend/database deployment, also fixed some weird import issues
