+++ apps/backend/api/index.ts
@@ -1,0 +1,3 @@
+import app from '../src/app.js';
+
+export default app;
