+++ apps/backend/src/middleware/errorHandler.ts
@@ -1,5 +1,5 @@
 import {Request, Response, NextFunction} from "express";
-import { errorResponse } from "../models/responseModel";
+import { errorResponse } from "../models/responseModel.js";
 
 interface ExtendedError extends Error {
     code?: string;
